requirements_training.txt
datasets
fastchat
httpx
huggingface-hub==0.34.4
numpy
openai
pandas
Pillow
python-dotenv
PyYAML
requests
tqdm
transformers==4.55.4
codetiming
hydra-core==1.3.2
peft
torch
torchvision
ray
pybind11
pylatexenc
tensorboard
tensordict>=0.8.0,<=0.9.1,!=0.9.0
torchdata
cachetools
fastapi
uvicorn
qwen_vl_utils
# extra dependencies based on arm_agent_requirements.txt
# core deep learning
accelerate==1.10.0
wandb
torchaudio
xformers==0.0.31
triton==3.3.1
timm==1.0.19
opencv-python
opencv-python-headless
scikit-image
ultralytics
onnxruntime
safetensors
sentencepiece
tokenizers
scipy
scikit-learn
seaborn
matplotlib
polars
pyarrow
sentence-transformers
gguf
gradio
gradio-client
flask
gunicorn
uvloop
chromadb
outlines
ipython
ipykernel
ipdb
ruff
py-spy
# this two may need to be installed manually
vllm==0.10.1.1
flash-attn==2.8.2